1682 fix init test (#1738)
* Fix init test

* Update sdk version
