+++ cmd/crates/soroban-test/tests/it/init.rs
@@ -1,14 +1,13 @@
 use assert_fs::prelude::*;
 use predicates::prelude::predicate;
-use soroban_test::{AssertExt, TestEnv};
+use soroban_test::TestEnv;
 
 #[test]
 fn init() {
     let sandbox = TestEnv::default();
-    let major = soroban_cli::commands::version::pkg()
-        .split('.')
-        .next()
-        .unwrap();
+    let cli_version = soroban_cli::commands::version::pkg();
+    let major = cli_version.split('.').next().unwrap();
+    let is_rc = cli_version.contains("rc");
     sandbox
         .new_assert_cmd("contract")
         .arg("init")
@@ -20,68 +19,12 @@ fn init() {
         .child("Cargo.toml")
         .assert(predicate::function(|c: &str| {
             let table = toml::from_str::<toml::Table>(c).unwrap();
-            table["workspace"]["dependencies"]["soroban-sdk"].as_str()
-                == Some(&format!("{major}.0.0"))
+            let sdk_version = table["workspace"]["dependencies"]["soroban-sdk"].as_str();
+            println!("Check expected version {major}.0.0 matches template's {sdk_version:?}");
+            if is_rc {
+                sdk_version.and_then(|x| x.split('-').next()) == Some(&format!("{major}.0.0"))
+            } else {
+                sdk_version == Some(&format!("{major}.0.0"))
+            }
         }));
-}
-
-#[test]
-fn init_and_deploy() {
-    let name = "hello_world";
-    let sandbox = TestEnv::default();
-
-    sandbox
-        .new_assert_cmd("contract")
-        .arg("init")
-        .arg("--name")
-        .arg(name)
-        .arg("project")
-        .assert()
-        .success();
-
-    let manifest_path = sandbox
-        .dir()
-        .join(format!("project/contracts/{name}/Cargo.toml"));
-    assert!(manifest_path.exists());
-
-    sandbox
-        .new_assert_cmd("contract")
-        .arg("build")
-        .arg("--manifest-path")
-        .arg(manifest_path)
-        .assert()
-        .success();
-
-    let target_dir = sandbox
-        .dir()
-        .join("project/target/wasm32-unknown-unknown/release");
-    assert!(target_dir.exists());
-
-    let assert = sandbox
-        .new_assert_cmd("contract")
-        .arg("deploy")
-        .arg("--wasm")
-        .arg(target_dir.join(format!("{name}.wasm")))
-        .assert();
-
-    let contract = assert.stdout_as_str();
-
-    assert.success();
-
-    let assert = sandbox
-        .new_assert_cmd("contract")
-        .arg("invoke")
-        .arg("--id")
-        .arg(contract)
-        .arg("--")
-        .arg("hello")
-        .arg("--to")
-        .arg("bar")
-        .assert();
-
-    let output = assert.stdout_as_str();
-
-    assert_eq!(output, r#"["Hello","bar"]"#);
-
-    assert.success();
 }
