+++ cmd/crates/soroban-test/tests/it/integration.rs
@@ -4,6 +4,7 @@ mod cookbook;
 mod custom_types;
 mod dotenv;
 mod hello_world;
+mod init;
 mod keys;
 mod snapshot;
 mod tx;
